Assembly in parallel
